type casting video to float

+++ codes/EDVR_inference.py
@@ -47,7 +47,7 @@ class EDVRWrapper:
         with torch.no_grad():
             # read LQ images
             imgs_LQ, _, info = torchvision.io.read_video(input_path)  # imgs_LQ: Tensor[T,H,W,C]
-            imgs_LQ = imgs_LQ.permute(0, 3, 1, 2).contiguous() / 255.0  # imgs_LQ: Tensor[T,C,H,W]
+            imgs_LQ = imgs_LQ.permute(0, 3, 1, 2).contiguous().float() / 255.0  # imgs_LQ: Tensor[T,C,H,W]
             max_idx = imgs_LQ.shape[0]
             input_batch = list()
             output_tensor = list()
